# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: WotBTools HTTP Contract
  version: 1.0.0
  description: FE ↔ BE wire contract. Domain models and Control/Worker contracts are separate authorities.
paths:
  /api/observability/client-events:
    post:
      operationId: reportClientEvent
      summary: Report a bounded allowlisted critical client failure
      description: Single event, maximum 4096 bytes. No client identity, free text, payload or credentials. Verified JWT sub alone supplies userId. Maximum 10 reports per minute per authenticated user and 300 globally; ingress separately limits anonymous client IPs.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientEvent' }
      responses:
        '204': { description: Accepted }
        '400': { description: Invalid event, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '413': { description: Body too large, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '429': { description: Rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
  /api/ai/reviews:
    post:
      operationId: createAiReview
      summary: Stream an AI replay review
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiReviewRequestV1' }
      responses:
        '200':
          description: SSE stream. Stage events have empty payloads; call2_token carries delta; done carries the review result; error carries a stable error code.
          content:
            text/event-stream:
              schema: { type: string }
          x-sse-events:
            call1_start: { $ref: '#/components/schemas/AiReviewStageEventPayload' }
            call1_done: { $ref: '#/components/schemas/AiReviewStageEventPayload' }
            evidence_done: { $ref: '#/components/schemas/AiReviewStageEventPayload' }
            call2_token: { $ref: '#/components/schemas/AiReviewTokenEventPayload' }
            done: { $ref: '#/components/schemas/AiReviewDonePayload' }
            error: { $ref: '#/components/schemas/AiReviewErrorEventPayload' }
        '400':
          description: Invalid request envelope or unsupported schema/locale
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: AI review role required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '409': { description: Duplicate correlation ID, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '413': { description: Request exceeds the 16 MiB limit, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '422': { description: Unsupported battle category, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '503': { description: Bounded AI worker queue is full, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /api/ai/reviews/{correlationId}/cancel:
    post:
      operationId: cancelAiReview
      summary: Cancel an in-flight AI review
      security: [{ bearerAuth: [] }]
      parameters:
        - name: correlationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Cancellation accepted }
        '400': { description: Invalid correlation ID, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '401': { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: AI review role required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: No active review for this correlation ID, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ClientEvent:
      type: object
      additionalProperties: false
      required: [event, platform, errorCode]
      properties:
        event: { type: string, enum: [client.bootstrap_failed, client.wasm_load_failed, client.android_webview_failed] }
        platform: { type: string, enum: [web, android] }
        errorCode: { type: string, enum: [CLIENT_BOOTSTRAP_FAILED, CLIENT_WASM_LOAD_FAILED, CLIENT_WEBVIEW_FAILED] }
        correlationId: { type: string, format: uuid }
...
    AiReviewRequestV1:
      type: object
      additionalProperties: false
      required: [schemaVersion, locale, correlationId, battle, reconstruction]
      properties:
        schemaVersion: { type: integer, const: 1 }
        locale: { type: string, enum: [zh-CN, en-US, ru-RU] }
        correlationId: { type: string, format: uuid }
        battle: { $ref: '#/components/schemas/AiReviewBattle' }
        reconstruction: { $ref: '#/components/schemas/AiReviewReconstruction' }
    AiReviewBattle:
      type: object
      additionalProperties: false
      required: [players]
      properties:
        arenaId: { type: [string, 'null'] }
        winnerTeam: { type: [integer, 'null'] }
        arenaBonusType: { type: [integer, 'null'] }
        version: { type: string }
        mapName: { type: string }
        durationS: { type: [number, 'null'] }
        startTime: { type: [integer, 'null'], format: int64 }
        settlementStartTime: { type: [integer, 'null'], format: int64 }
        settlementFinishReasonRaw: { type: [integer, 'null'] }
        settlementDurationSec: { type: [number, 'null'] }
        recorder: { type: string }
        recorderVehicle: { type: string }
        clientVersion: { type: string }
        rosterComplete: { type: [boolean, 'null'] }
        players: { type: array, items: { $ref: '#/components/schemas/AiReviewPlayerResult' } }
    AiReviewPlayerResult:
      type: object
      additionalProperties: false
      required: [accountId, team, tankId, nickname, survived]
      properties:
        accountId: { type: integer, format: int64 }
        team: { type: integer }
        tankId: { type: integer, format: int64 }
        nShots: { type: integer }
        nHitsDealt: { type: integer }
        nPenetrationsDealt: { type: integer }
        damageDealt: { type: integer }
        damageAssisted: { type: integer }
        damageReceived: { type: integer }
        nHitsReceived: { type: integer }
        nPenetrationsReceived: { type: integer }
        nEnemiesDamaged: { type: integer }
        kills: { type: integer }
        damageBlocked: { type: integer }
        victoryPointsEarned: { type: integer }
        victoryPointsSeized: { type: integer }
        survived: { type: boolean }
        xp: { type: integer }
        credits: { type: integer }
        nickname: { type: string }
        clan: { type: string }
        prebattleGroupId: { type: [integer, 'null'], format: int64 }
        rank: { type: [integer, 'null'], format: int64 }
        tankName: { type: string }
        contribution: { type: [number, 'null'] }
        kast: { type: [number, 'null'] }
        impact: { type: [number, 'null'] }
        observedMaxHp: { type: [integer, 'null'] }
        entryHpSource: { type: [string, 'null'] }
        entryHp: { type: [integer, 'null'] }
        tankTier: { type: [string, integer, 'null'] }
        tankType: { type: string }
        tankNation: { type: string }
        alphaDamage: { type: [string, integer, 'null'] }
        deathTimeMillis: { type: integer, format: int64 }
        settlementResultEntityId: { type: integer, format: int64 }
        settlementLifeTimeSec: { type: number }
        settlementKillerResultEntityId: { type: [integer, 'null'], format: int64 }
        settlementDeathReasonRaw: { type: [integer, 'null'] }
        killerAccountId: { type: [integer, 'null'], format: int64 }
        survivalTimeSec: { type: number }
        raw: { type: [object, 'null'], additionalProperties: { type: array, items: {} } }
    AiReviewReconstruction:
      type: object
      additionalProperties: false
      required: [participants, events, coverage]
      properties:
        battleDurationSec: { type: number }
        battleStartRawClockSec: { type: [number, 'null'] }
        participants: { type: array, items: { $ref: '#/components/schemas/AiReviewParticipant' } }
        events: { type: array, items: { $ref: '#/components/schemas/AiReviewEvent' } }
        checkpoints: { type: [array, 'null'], items: { type: object, additionalProperties: true } }
        finalState: { type: [object, 'null'], additionalProperties: true }
        coverage: { $ref: '#/components/schemas/AiReviewCoverage' }
    AiReviewParticipant:
      type: object
      additionalProperties: false
      required: [accountId, nickname, team, tankId, tankCode, recorder]
      properties:
        accountId: { type: integer, format: int64 }
        nickname: { type: string }
        team: { type: integer }
        tankId: { type: integer }
        tankCode: { type: [string, 'null'] }
        recorder: { type: boolean }
    AiReviewEvent:
      type: object
      required: [type, sequence]
      properties:
        type:
          type: string
          enum: [AimRayStateEvent, AmmunitionSelectionChangedEvent, AmmunitionStateEvent, ArenaPeriodChangedEvent, AttachedTransformEvent, ConsumableLifecycleEvent, DamageEvent, EntityAuxiliaryBlobEvent, EntityCreatedEvent, EntityRemovedEvent, GunMarkerSizeEvent, HealthChangedEvent, MaterializationAnnouncedEvent, MaterializationEvent, ParticipantMappingEvent, PositionChangedEvent, ProjectileLaunchedEvent, ProjectileResolutionEvent, ProjectileTerminalEvent, RawSupremacyBaseUpdate, RecorderHealthChangedEvent, ReplayStreamClosedEvent, RoundFinishedEvent, SessionDecisecondLowByteEvent, ShotResultEvent, SupremacyBaseStateTransition, SupremacyPointsChangedEvent, TargetingInfoSnapshotEvent, TurretDirectionChangedEvent, UnknownReplayEvent, UnsupportedDamageEvent, VehicleDestroyedEvent, VehicleFiredEvent, VehicleHealthStateEvent, VehicleHitEvent, VehicleModuleCrewStateEvent, VehicleVehicleCollisionEvent]
        sequence: { type: integer }
      additionalProperties: true
    AiReviewCoverage:
      type: object
      additionalProperties: false
      required: [totalPackets, decodedPackets, partiallyDecodedPackets, unknownPackets, failedPackets, decodedPacketRatio, packetTypes]
      properties:
        totalPackets: { type: integer }
        decodedPackets: { type: integer }
        partiallyDecodedPackets: { type: integer }
        unknownPackets: { type: integer }
        failedPackets: { type: integer }
        decodedPacketRatio: { type: number }
        packetTypes: { type: object, additionalProperties: { type: object, additionalProperties: true } }
    AiReviewStageEventPayload:
      type: object
      additionalProperties: false
    AiReviewTokenEventPayload:
      type: object
      additionalProperties: false
      required: [delta]
      properties:
        delta: { type: string }
    AiReviewErrorEventPayload:
      type: object
      additionalProperties: false
      required: [id, errorCode]
      properties:
        id: { type: [string, 'null'] }
        errorCode: { $ref: '#/components/schemas/AiReviewErrorCode' }
    AiReviewErrorCode:
      type: string
      enum: [AI_CANCELLED, AI_NOT_CONFIGURED, AI_INVALID_REQUEST, AI_TIMEOUT, AI_UPSTREAM_UNAVAILABLE, AI_AUTHENTICATION_ERROR, AI_RATE_LIMITED, AI_CONTEXT_TOO_LARGE, AI_EMPTY_RESPONSE, AI_RESPONSE_INVALID, AI_REVIEW_SCHEMA_FAILED, AI_REVIEW_GROUNDING_FAILED, AI_TIMELINE_UNUSABLE, AI_PROMPT_MANDATORY_SECTION_TOO_LARGE]
    TeamAiReviewSummary:
      type: object
      additionalProperties: false
      required: [verdict, primaryDiagnosis]
      properties:
        verdict: { type: string, minLength: 1, maxLength: 4000 }
        primaryDiagnosis: { type: string, minLength: 1, maxLength: 4000 }
    TeamAiReviewEpisode:
      type: object
      additionalProperties: false
      required: [id, startSec, endSec, title, analysis, playerKeys]
      properties:
        id: { type: string, minLength: 1, maxLength: 64 }
        startSec: { type: [integer, 'null'], minimum: 0 }
        endSec: { type: [integer, 'null'], minimum: 0 }
        title: { type: string, minLength: 1, maxLength: 240 }
        analysis: { type: string, minLength: 1, maxLength: 8000 }
        playerKeys:
          type: array
          maxItems: 8
          items: { type: string, minLength: 1, maxLength: 64 }
    TeamAiTrainingSuggestion:
      type: object
      additionalProperties: false
      required: [title, content, episodeId]
      properties:
        title: { type: string, minLength: 1, maxLength: 240 }
        content: { type: string, minLength: 1, maxLength: 6000 }
        episodeId: { type: [string, 'null'], maxLength: 64 }
    TeamAiReviewFocus:
      type: object
      additionalProperties: false
      required: [playerKey, episodeId, reason]
      properties:
        playerKey: { type: string, minLength: 1, maxLength: 64 }
        episodeId: { type: string, minLength: 1, maxLength: 64 }
        reason: { type: string, minLength: 1, maxLength: 2000 }
    TeamAiHighContributor:
      type: object
      additionalProperties: false
      required: [playerKey, episodeId, reason]
      properties:
        playerKey: { type: string, minLength: 1, maxLength: 64 }
        episodeId: { type: string, minLength: 1, maxLength: 64 }
        reason: { type: string, minLength: 1, maxLength: 2000 }
    TeamAiPlayerIdentity:
      type: object
      additionalProperties: false
      required: [playerKey, displayName, tankName]
      properties:
        playerKey: { type: string, minLength: 1, maxLength: 64 }
        displayName: { type: string, maxLength: 240 }
        tankName: { type: string, maxLength: 240 }
    TeamAiReviewResult:
      type: object
      additionalProperties: false
      required: [summary, episodes, trainingSuggestions, reviewFocus, highContributors]
      properties:
        summary: { $ref: '#/components/schemas/TeamAiReviewSummary' }
        episodes:
          type: array
          maxItems: 6
          items: { $ref: '#/components/schemas/TeamAiReviewEpisode' }
        trainingSuggestions:
          type: array
          maxItems: 12
          items: { $ref: '#/components/schemas/TeamAiTrainingSuggestion' }
        reviewFocus:
          type: array
          maxItems: 2
          items: { $ref: '#/components/schemas/TeamAiReviewFocus' }
        highContributors:
          type: array
          maxItems: 2
          items: { $ref: '#/components/schemas/TeamAiHighContributor' }
    AiReviewDonePayload:
      type: object
      additionalProperties: false
      required: [analysis, preBattleSection, capability, teamReview, teamPlayers]
      properties:
        analysis: { type: [string, 'null'] }
        preBattleSection: { type: [string, 'null'] }
        capability: { type: [string, 'null'], enum: [AVAILABLE, AVAILABLE_WITH_LIMITED_TIMELINE, UNAVAILABLE, null] }
        teamReview: { oneOf: [{ $ref: '#/components/schemas/TeamAiReviewResult' }, { type: 'null' }] }
        teamPlayers:
          type: array
          items: { $ref: '#/components/schemas/TeamAiPlayerIdentity' }
...
    ApiError:
      type: object
      additionalProperties: false
      required: [id, errorCode, errorMsg, status, retryable, details, timestamp]
      properties:
        id: { type: [string, 'null'] }
        errorCode:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          description: Stable infrastructure code or a legacy uppercase domain code during the migration boundary.
        errorMsg: { type: [string, 'null'] }
        status: { type: integer, minimum: 400, maximum: 599 }
        retryable: { type: boolean }
        details: { type: object, additionalProperties: true }
        timestamp: { type: [string, 'null'], format: date-time }
...